+++ src/index.js
@@ -1429,7 +1429,7 @@ router.post('/:path/setting', async request => {
     try {
         if (request.headers.get('Content-Type') === 'application/json') {
             const cookie = Cookies.parse(request.headers.get('Cookie') || '')
-            const { mode, share, theme, width, shareFont, previewDevice, publicIndex } = await request.json()
+            const { mode, share, theme, width, shareFont, previewDevice, splitDirection, publicIndex } = await request.json()
 
             const { value, metadata } = await queryNote(path)
             const { valid } = await checkAuth(cookie, path)
@@ -1444,6 +1444,7 @@ router.post('/:path/setting', async request => {
                         ...width !== undefined && { width },
                         ...shareFont !== undefined && { shareFont },
                         ...previewDevice !== undefined && { previewDevice },
+                        ...splitDirection !== undefined && { splitDirection },
                         ...publicIndex !== undefined && { publicIndex: publicIndex === true },
                     }